Add integration testing code
The `dmsa test` command takes a URL option and tests the web service
endpoints at that URL for '200 OK' status codes. If the service is in
debug mode, the testing code attempts to extract and print the relevant
traceback when a URL fails. In order to support this, the MODELS and
DIALECTS constants are moved into the settings file and the URLs are
included in those dicts, which required some changes to the template
code.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>