Deploy by branch instead of by sha
This should reduce the number of AWS environments that are created,
which should reduce the frequency with which we hit the 5 EIP limit.

Also, configure git username before trying to tag a release.

Also, use the newly-standard docker 1.6.2 service instead of manually
installing it.

Also, format the jq install more cleanly (to my taste).

Bump version to 0.4.5-alpha

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>