Make dmsa installable
Fixes #43

This is not the best way to fix the problem, but the only way that
preserves the dmsa.make_model_from_service API. I moved all the
sqlalchemy imports inside of functions in the makers module.

Also, the Dockerfile and circle.yml files are updated with simpler
install steps that are now possible.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>