Add requirements.txt to MANIFEST.in
This fixes a problem with tox. Tox tries to create a source
distribution, probably to install into the virtual environment it
creates, and the lack of requirements.txt was preventing this.

There are other files missing from the source distribution, but
presumably these don't matter for normal Python use. E.g.:

.coveragerc
.dockerignore
Dockerfile
tox.ini
lib/*.zip